Extract SseReader from client transport
